Some refactoring of settings activity, but we should just make a new ticket for that beast.
