+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         <service android:name=".Services.SchedulingService" />
 
         <activity
-            android:name=".SettingsActivity"
+            android:name=".Activities.SettingsActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar.FullScreen">
